+++ src/ssh/HISTORY.md
@@ -1,10 +1,18 @@
 Release History
 ===============
-0.1.7
+0.1.9
 -----
 * Add support for connecting to Arc Servers using AAD issued certificates.
 * Add support for connecting to local users on Azure VMs and Arc Server using certs-based, key-based, and password-based authentication
-* Add --ssh-client-path, --resource-id, --local-user, --cert-file, and --delete-privkey arguments.
+* Add --ssh-client-path, --resource-id, --local-user, --cert-file arguments.
+
+0.1.8
+-----
+* Rollback from version 0.1.7 to 0.1.6 to remove preview features.
+
+0.1.7
+-----
+* Introduced preview features.
 
 0.1.6
 -----